+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Expand to screen button which allows you to display UI on your screen.
+- Added modal frame that can be used to render top level elements like dropdown list.
+- Fixed story tree item's texts being too big.
 
 ### Changed
 - When a studio test is initiated, HoarceKat will no longer display an update message.